Enhanced CallerId:
The CallerId, that gets written to the database, consists
on the first user phone number. For an easy
way to display the callers name in the phones display,

we want to save the first- and the lastname in the same
field, now.

The CallerId length is 80. To garantee, that
the whole number is always stored in this field we
strip the amount of chars from the left side of the string,
that would exceeds the maximum 80 chars limit.



git-svn-id: https://oss.gonicus.de/repositories/gosa/trunk@21277 594d385d-05f5-0310-b6e9-bd551577e9d8